Fix britghness calculation bug. Implement reading smoothing (Not working)

+++ .ccls-cache/@home@nongtajkrub@Coding@Ard@Helios/src@func.hpp
@@ -3,10 +3,10 @@
 #include <stdint.h>
 #include <math.h>
 
-inline uint8_t min(uint8_t min, uint8_t val) {
+inline uint8_t min(uint16_t min, uint16_t val) {
 	return (val < min) ? min : val;
 }
 
-inline uint8_t max(uint8_t max, uint8_t val) {
+inline uint8_t max(uint16_t max, uint16_t val) {
 	return (val > max) ? max : val;
 }
